refactor(frontend): move component config out of components
