+++ apps/desktop/src/config/settings-navigation.ts
